requirements.txt
requests
tabula-py
pandas
matplotlib
schedule